Add comment crud; fixed: types, post crud

+++ back/src/post/post.service.ts
@@ -1,9 +1,11 @@
-import { Injectable } from '@nestjs/common';
-import { CreatePostDto } from './dto/create-post.dto';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PostRepository } from './post.repository';
 import { CreatePost } from './types/create-post.interface';
-import { UpdatePost } from './types/update-post.interface';
 import { findOptions } from './types/find-options.type';
 
 @Injectable()
@@ -21,11 +23,26 @@ export class PostService {
     return await this.postRepository.getOne(id);
   }
 
-  async update(id: string, updatePostDto: UpdatePostDto) {
+  async update(id: string, updatePostDto: UpdatePostDto, userId: string) {
+    const post = await this.findOne(id);
+    if (!(post.userId === userId)) {
+      throw new ForbiddenException();
+    }
     return await this.postRepository.update(id, updatePostDto);
   }
 
-  async remove(id: string) {
-    return await this.postRepository.delete(id);
+  async remove(id: string, userId: string) {
+    try {
+      const exPost = await this.findOne(id);
+      if (!exPost) {
+        throw new NotFoundException(`Post with id ${id} not found`);
+      }
+      if (exPost.userId !== userId) {
+        throw new ForbiddenException();
+      }
+      return await this.postRepository.delete(id);
+    } catch (err) {
+      return err;
+    }
   }
 }
